refactored makefile

+++ Makefile
@@ -1,27 +1,38 @@
-CPP		:= g++
+CPP				:= g++
 
-SRC		:= src
-BIN		:= build
-EXE		:= $(BIN)/main
-OBJ		:= $(BIN)/obj
-DEPS	:= $(BIN)/deps
+SRC				:= ./src
+BIN				:= ./build
+EXE				:= $(BIN)/main
+OBJ				:= $(BIN)/obj
+DEPS			:= $(BIN)/deps
 
-SRCS	:= $(shell find . -name '*.cpp')
-VPATH	:= $(dir $(SRCS))
-OBJS	:= $(patsubst %.cpp, $(OBJ)/%.o, $(notdir $(SRCS)))
-INCLUDE	:= $(shell find $(SRC) -type d | xargs printf '\-I%s ' | xargs)
+SRCS			:= $(shell find . -name '*.cpp')
+VPATH			:= $(dir $(SRCS))
+OBJS			:= $(patsubst %.cpp, $(OBJ)/%.o, $(notdir $(SRCS)))
+INCLUDE		:= $(shell find $(SRC) -type d | xargs printf '\-I%s ' | xargs)
 
-CXXFLAGS += -std=c++17 -pedantic -Wall -O3 $(INCLUDE)
-LDFLAGS += -lm
+CXXFLAGS 	+= -std=c++1z -pedantic -Wall -O3 $(INCLUDE)
+LDFLAGS  	+= -lm
+
+MULTILINE ?= 0
 
 ifndef ECHO
-HIT_N			?= 0
-HIT_TOTAL := $(shell ECHO="HIT_MARK" $(MAKE) $(MAKECMDGOALS) --dry-run | grep -c "HIT_MARK")
-HIT_COUNT = $(eval HIT_N = $(shell expr $(HIT_N) + 1))$(HIT_N)
-ECHO = echo "[$(HIT_COUNT)/$(HIT_TOTAL)]"
+HIT_N     ?= 0
+HIT_TOTAL	 = $(eval HIT_TOTAL := $$(shell ECHO="HIT_MARK" $(MAKE) $(MAKECMDGOALS) --dry-run | grep -c "HIT_MARK"))$(HIT_TOTAL)
+HIT_COUNT  = $(eval HIT_N = $(shell expr $(HIT_N) + 1))$(HIT_N)
+
+ifneq ($(MULTILINE),0)
+EOL       := \\n
+else
+RESET     := \033[2K\r
 endif
 
-.PHONY: all run build rebuild clean
+ENDLINE    = $(shell [[ $(HIT_N) == $(HIT_TOTAL) ]] && echo '$(RESET)$(MAKECMDGOALS) successful\\n')
+
+ECHO       = printf "$(RESET)[$(HIT_COUNT)/$(HIT_TOTAL)] %s$(EOL)$(ENDLINE)"
+endif
+
+.PHONY: all run exec build rebuild clear remove
 .SILENT:
 
 all: clean build
@@ -29,9 +40,9 @@ all: clean build
 
 build: $(EXE)
 
-rebuild: clean build
+rebuild: clean $(EXE)
 
-run: build
+exec run: build
 	$(EXE)
 
 $(EXE): $(OBJS) | $(OBJ)
@@ -42,11 +53,14 @@ $(OBJ)/%.o: %.cpp | $(OBJ) $(DEPS)
 	$(ECHO) "building $@..."
 	$(CPP) $(CXXFLAGS) -c -MMD -MF $(DEPS)/$(shell echo $@ | xargs -L 1 basename | sed 's/o$$/d/') $< -o $@
 
-$(OBJ) $(DEPS):
+$(BIN) $(OBJ) $(DEPS):
 	mkdir -p $@
 
-clean:
+clean: remove
 	$(ECHO) "cleaning..."
-	rm -rf $(BIN)
+	[ ! -d $(BIN)_ ] || rm -rf $(BIN)_
+
+remove:
+	[ ! -d $(BIN) ] || mv $(BIN) $(BIN)_
 
 -include $(DEPS)/*.d
